Feature/steering sensor

+++ src/fsd_rl/src/fsd_env.py
@@ -23,7 +23,7 @@ from fssim_common.msg import ResState, State
 
 from sensor_msgs import point_cloud2
 from sensor_msgs.msg import PointCloud2
-from std_msgs.msg import Int16, Int32
+from std_msgs.msg import Int16, Int32, Float32
 
 #from vesc_msgs.msg import *
 
@@ -40,40 +40,42 @@ THROTTLE_START = float(0.08)  # Starting throttle position (for initial accelera
 THROTTLE_START_TIME = 3.0  # Time to accelerate for
 THROTTLE_SET = float(0.037)  # Set throttle position
 NUM_CONES = 3  # Number of cones of each colour stored and used
-RANGE = 10  # Range of cameras
-STEER_ANG_MIN = -1  # Try change to 0.4 & -0.4
-STEER_ANG_MAX = 1
+RANGE = 10  # Range of cameras (note that range is set in sensors_1.yaml)
+STEER_ANG_DEG_LIMIT = 45 # Estimate of steering angle limit in sim (at +1/-1)
+STEER_ANG_MIN = -0.4
+STEER_ANG_MAX = 0.4
+STEER_ANG_RATE_MAX = 67.5 # Deg/s
 
 #IDENT_BLUE = 1  # Identifiers for blue and yellow cones
 #IDENT_YELLOW = -1
 
 class FsdEnv(gym.Env):
 
-    def __init__(self):
-        # Set up simulated environment
-        # Subscribers
-        rospy.Subscriber('/control/pure_pursuit/control_command', ControlCommand, self.callback_cmd)  # Control command
-        rospy.Subscriber('/fssim/res_state', ResState, self.callback_res_state)  # Vehicle OK?
-        rospy.Subscriber('/fssim/stat_lap_count', Int16, self.callback_lap)  # Lap counter
-        rospy.Subscriber('/camera/cones', PointCloud2, self.callback_cones)  # Cone locations
+    def __init__(self, sim=True):
+        self.sim_true = sim  # True if simulated, false if physical
 
-        # TODO: Complete subscribers and publishers for physical
-        # Subscribers
-        # rospy.Subscriber('aruco_detector', ,self.callback_cones_physical) # Cone positions - SERVICE
-        # rospy.Subscriber('/control/physical/steering_ang', , self.callback_steering_physical) # Steering angle (Need linked/timed service??)
+        # Publishers and subscribers
+        if self.sim_true:
+            print('SIM')
+            rospy.Subscriber('/control/pure_pursuit/control_command', ControlCommand, self.callback_cmd)  # Control command
+            rospy.Subscriber('/fssim/res_state', ResState, self.callback_res_state)  # Vehicle OK?
+            rospy.Subscriber('/fssim/stat_lap_count', Int16, self.callback_lap)  # Lap counter
+            rospy.Subscriber('/camera/cones', PointCloud2, self.callback_cones)  # Cone locations
 
-        # Publishers
-        # Control command
-        self.cmd_airl = rospy.Publisher('/control/pure_pursuit/control_command', ControlCommand, queue_size=5)
-        #self.cmd_phys = rospy.Publisher('/control/pure_pursuit/control_command', ControlCommand, queue_size=5)
+            self.cmd_airl = rospy.Publisher('/control/pure_pursuit/control_command', ControlCommand, queue_size=5)
+        else:
+            print('Phys')
+            # rospy.Subscriber('aruco_detector', ,self.callback_cones_phys) # Cone positions - SERVICE
+            rospy.Subscriber('/physical/steering/norm_ang', Float32, self.callback_cmd_phys)  # Steering angle (Need linked/timed service??)
+
+            self.cmd_phys = rospy.Publisher('/control/steering/norm_ang', Float32, queue_size=5)
 
         # Define observation and action spaces
         # Observation space
-        # Note that range is set in sensors_1.yaml
         self.x_low = float(0)
-        self.x_high = float(10)
-        self.y_low = float(-5)
-        self.y_high = float(5)
+        self.x_high = float(RANGE)
+        self.y_low = -float(RANGE)/2.0
+        self.y_high = float(RANGE)/2.0
 
         cones_x_low = np.full((2*NUM_CONES, 1), self.x_low)
         cones_x_high = np.full((2*NUM_CONES, 1), self.x_high)
@@ -82,8 +84,8 @@ class FsdEnv(gym.Env):
         #cones_col_low = np.full((2*NUM_CONES, 1), min(IDENT_BLUE, IDENT_YELLOW))
         #cones_col_high = np.full((2*NUM_CONES, 1), max(IDENT_BLUE, IDENT_YELLOW))
 
-        cones_low = np.concatenate((cones_x_low, cones_y_low), axis=1)
-        cones_high = np.concatenate((cones_x_high, cones_y_high), axis=1)
+        cones_low = np.concatenate((cones_x_low, cones_y_low), axis=1) #, cones_col_low
+        cones_high = np.concatenate((cones_x_high, cones_y_high), axis=1) #, cones_col_high
 
         # Flatten cones array
         cones_low = cones_low.flatten()
@@ -92,14 +94,15 @@ class FsdEnv(gym.Env):
         self.observation_space = spaces.Box(low=cones_low, high=cones_high, dtype=float)
 
         # Action space
-        self.action_space = spaces.Box(low=float(STEER_ANG_MIN), high=float(STEER_ANG_MAX), shape=(1,), dtype=float)
+        #self.action_space = spaces.Box(low=float(STEER_ANG_MIN), high=float(STEER_ANG_MAX), shape=(1,), dtype=float)
 
         # Establish connection with simulator
         self.gazebo = GazeboConnection()
         rospy.set_param('/use_sim_time', 'true')
 
         # gets training parameters from param server
-        self.running_step = rospy.get_param("/running_step")
+        self.running_step = rospy.get_param("/running_step") # Step size/sample time
+        self.rate_limit_sample = (STEER_ANG_RATE_MAX / STEER_ANG_DEG_LIMIT) * self.running_step # Maximum normalised ang rate (norm ang/sample)
 
         # Set tracking instance variables
         self.np_random = self.cnt_step = self.cnt_lap = self.shutdown = 0  # Tracking variables
@@ -109,30 +112,53 @@ class FsdEnv(gym.Env):
         self.obs_cmd = self.obs_cones = None  # Last "observed" command and cone positions
 
         # Physical-only vars
-        self.phys_sa = self.phys_cones = None # Last "observed" command and cone positions
+        self.phys_sa = self.phys_cones = None  # Last "observed" command and cone positions
 
         self.helper_reset_vars()
+
+    ### DYNAMIC PROPERTIES ###
+    # Change action space on every step for rate limiting
+    @property
+    def action_space(self):
+        # Rate limiter
+        # Set limits for rate
+        max_rate = self.observation_prev["steering_angle"] + self.rate_limit_sample
+        min_rate = self.observation_prev["steering_angle"] - self.rate_limit_sample
+
+        # Angle limiter
+        # Clip action to limits (required as mlp in 'policy' outputs action in -1 -> 1 space)
+        max_rate = np.clip(max_rate, STEER_ANG_MIN, STEER_ANG_MAX)
+        min_rate = np.clip(min_rate, STEER_ANG_MIN, STEER_ANG_MAX)
+
+        action_space = spaces.Box(low=float(min_rate), high=float(max_rate), shape=(1,), dtype=float)
+
+        return action_space
 
     ### CALLBACKS ###
     # Stores the current command sent
     def callback_cmd(self, data_cmd):
         self.obs_cmd = data_cmd
 
-    # TODO: Implement physical s.a. callback
-    def callback_phys_sa(self, data_sa):
-        # Process input to make equivalent to simulation
-
+    def callback_cmd_phys(self, data_sa):
         self.phys_sa = data_sa
+
+        temp = Float32()
+        r = rospy.Rate(10)
+        while True:
+            #print(self.phys_sa.data)
+            temp.data = self.phys_sa.data
+            #print(temp)
+            self.cmd_phys.publish(temp)
+            r.sleep()
 
     # Stores current cone locations
     def callback_cones(self, data_pt_cloud):
         self.obs_cones = list(point_cloud2.read_points(data_pt_cloud, skip_nans=True, field_names=("x", "y", "probability_blue", "probability_yellow", "probability_orange", "probability_other")))
 
     #TODO: Implement physical cones callback
-    def callback_phys_cones(self, data_cones):
+    def callback_cones_phys(self, data_cones):
         # Process input to make equivalent to simulation
-
-        self.phys_cones = data_cones
+        self.obs_cones = data_cones
 
     # Stores shutdown state
     def callback_res_state(self, data_res_state):
@@ -155,19 +181,21 @@ class FsdEnv(gym.Env):
     def reset(self):
         # 1st: resets the simulation to initial values
         # Kill and restart /automated_res
-        os.system("rosnode kill /automated_res")
-        rospy.wait_for_message('/fssim/res_state', ResState)
+        if self.sim_true:
+            os.system("rosnode kill /automated_res")
+            rospy.wait_for_message('/fssim/res_state', ResState)
 
-        # 2nd: Unpauses simulation
-        self.gazebo.unpauseSim()
+            # 2nd: Unpauses simulation
+            self.gazebo.unpauseSim()
 
-        # Would use reset helper here
+            # Would use reset helper here
 
         # 3rd: takes an observation of the initial condition of the robot
         observation = self.make_observation()
 
-        # 5th: pauses simulation
-        self.gazebo.pauseSim()
+        if self.sim_true:
+            # 5th: pauses simulation
+            self.gazebo.pauseSim()
 
         # Reset tracking vars and get initial state
         self.helper_reset_vars()
@@ -177,31 +205,54 @@ class FsdEnv(gym.Env):
 
     # Performs a single step in the environment
     def step(self, action):
+        # Angle limiter
+        # Clip action to limits (required as mlp in 'policy' outputs action in -1 -> 1 space)
+        action = np.clip(action, self.action_space.low, self.action_space.high)
+        #
+        # # Rate limiter
+        # # Check if rate limiting required
+        # if abs(action-self.observation_prev["steering_angle"]) > self.rate_limit_sample:
+        #     if action > self.observation_prev["steering_angle"]:
+        #         action = self.observation_prev["steering_angle"] + self.rate_limit_sample
+        #     else:
+        #         action = self.observation_prev["steering_angle"] - self.rate_limit_sample
+
         # Given the action selected by the learning algorithm,
         # we perform the corresponding movement of the robot
 
-        # 1. Set control command (normalised between -1 and 1)
-        next_action = ControlCommand()
-        next_action.steering_angle.data = action  # Set steering angle to input steering angle
+        if self.sim_true: # Simulation step
+            # 1. Set control command (normalised between -1 and 1)
+            next_action = ControlCommand()
+            next_action.steering_angle.data = action  # Set steering angle to input steering angle
 
-        # 2. Then we send the command to the robot and let it go
-        # for running_step seconds
-        self.gazebo.unpauseSim()
+            # 2. Then we send the command to the robot and let it go
+            # for running_step seconds
+            self.gazebo.unpauseSim()
 
-        seconds_start_step = rospy.get_time()  # Change to get_rostime()??
+            seconds_start_step = rospy.get_time()  # Change to get_rostime()??
 
-        # Send command to sim for running time
-        while (rospy.get_time() - seconds_start_step) < self.running_step:
-            # Send higher throttle command when accelerating at start
-            if self.running_step*self.cnt_step < THROTTLE_START_TIME:
-                next_action.throttle.data = THROTTLE_START
-            else:
-                next_action.throttle.data = THROTTLE_SET
+            # Send command to sim for running time
+            while (rospy.get_time() - seconds_start_step) < self.running_step:
+                # Send higher throttle command when accelerating at start
+                if self.running_step*self.cnt_step < THROTTLE_START_TIME:
+                    next_action.throttle.data = THROTTLE_START
+                else:
+                    next_action.throttle.data = THROTTLE_SET
 
-            self.cmd_airl.publish(next_action)
+                self.cmd_airl.publish(next_action)
 
-        observation = self.make_observation()
-        self.gazebo.pauseSim()
+            observation = self.make_observation()
+            self.gazebo.pauseSim()
+
+        else:  # Step in physical world
+            next_action = Float32()
+            next_action.data = action
+
+            seconds_start_step = rospy.get_time()
+            while (rospy.get_time() - seconds_start_step) < self.running_step:
+                self.cmd_phys.publish(next_action)
+
+            observation = self.make_observation()
 
         # Get an evaluation based on what happened in the sim
         reward, done = self.process_data(observation, self.observation_prev)
@@ -210,7 +261,6 @@ class FsdEnv(gym.Env):
         # Update variables
         self.cnt_step += 1
         self.observation_prev = observation
-        #self.cnt_step_total += 1
 
         return state, reward, done, {}
 
@@ -219,7 +269,7 @@ class FsdEnv(gym.Env):
     def process_data(self, observation, observation_prev):
         # Check if simulation is done (if vehicle has left track)
         done = False
-        if self.shutdown == 1:
+        if self.shutdown == 1 or self.cnt_lap >= 2:
             done = True
 
         # Calculate reward
@@ -232,11 +282,15 @@ class FsdEnv(gym.Env):
     def make_observation(self):
         cones_blue = list()  # Left side - list of tuples
         cones_yellow = list()  # Right side - list of tuples
-        steering_angle = self.obs_cmd.steering_angle.data
+        if self.sim_true:
+            steering_angle = self.obs_cmd.steering_angle.data
+        else:
+            steering_angle = self.phys_sa.data
 
         i = 0
-        while i < len(self.obs_cones):
-            temp = self.obs_cones[i]
+        curr_obs_cones = self.obs_cones
+        while i < len(curr_obs_cones):
+            temp = curr_obs_cones[i]
 
             # Assign cones to the correct lists based on the highest probability
             if temp.index(max(temp[2:5])) == 2:  # Cone is blue
@@ -270,7 +324,9 @@ class FsdEnv(gym.Env):
         self.obs_cmd = ControlCommand()  # Latest control command
         self.obs_cones = list()  # Cone positions relative to car
         self.shutdown = 0  # Vehicle quit due to emergency
-        # self.init_pose = PoseWithCovarianceStamped()
+
+        self.phys_sa = Float32()
+        #self.phys_cones = None
 
         self.observation_prev = self.make_observation()
 
@@ -279,7 +335,6 @@ class FsdEnv(gym.Env):
         cones_blue_array = np.array(observation["cones_blue"])
         cones_yellow_array = np.array(observation["cones_yellow"])
         #cones_identifiers = np.concatenate((np.full((NUM_CONES, 1), IDENT_BLUE), np.full((NUM_CONES, 1), IDENT_YELLOW)), axis=0)
-        #cones_identifiers = np.concatenate(np.full((NUM_CONES, 1), np.full(NUM_CONES, 1)), axis=0)
 
         #observation = np.concatenate((np.concatenate((cones_blue_array, cones_yellow_array), axis=0), cones_identifiers), axis=1)
         observation = np.concatenate((cones_blue_array, cones_yellow_array), axis=0)
@@ -317,8 +372,6 @@ class FsdEnv(gym.Env):
 
         # Find distance to right set of cones
 
-
-
         return reward
 
     # Reward for moving towards centre between furthest cones
@@ -337,7 +390,8 @@ class FsdEnv(gym.Env):
                 dist_list = self.helper_find_dist([target])
                 dist_to_target = dist_list[0]
 
-                reward = RANGE/dist_to_target  # Reward based on inverse distance to target
+                #reward = max(RANGE/dist_to_target, RANGE/0.25)  # Reward based on inverse distance to target
+                reward = RANGE / dist_to_target
 
         return reward
